[ALPS08543716] audio: mt6991 afe driver porting
1. fix BT HFP no sound
2. fix clk related control
3. update ap dmic patch from rayas
4. update GASRC coeff to support 7.5k up
5. fix mtkaif calibration
6. support dram address > 4GB
7. add codec regulator error handle
8. refine scp hw control flow in mt6681

MTK-Commit-Id: 22080d7eed7b42dd5740c085d631f06b7a56d601

CR-Id: ALPS08543716
Feature: alsa_sound_soc

+++ sound/soc/mediatek/common/mtk-afe-platform-driver.h
@@ -32,6 +32,6 @@ void mtk_afe_pcm_free(struct snd_soc_component *component, struct snd_pcm *pcm);
 
 int mtk_afe_combine_sub_dai(struct mtk_base_afe *afe);
 int mtk_afe_add_sub_dai_control(struct snd_soc_component *component);
-unsigned int word_size_align(unsigned int in_size);
+unsigned long word_size_align(unsigned long in_size);
 #endif
